#1511 extract commons-server module
