+++ bootstrap-core/pom.xml
@@ -70,9 +70,6 @@
                             <filters>
                                 <filter>
                                     <artifact>com.navercorp.pinpoint:pinpoint-commons</artifact>
-                                    <excludes>
-                                        <exclude>com/navercorp/pinpoint/common/bo/**</exclude>
-                                    </excludes>
                                 </filter>
                             </filters>
                         </configuration>
