+++ commons-server/.gitignore
@@ -1,0 +1,5 @@
+/.settings/
+/*.iml
+/.project
+/target/
+/.classpath
