+++ commons-server/src/main/java/com/navercorp/pinpoint/common/bo/AgentInfoBo.java
