+++ commons-server/src/main/java/com/navercorp/pinpoint/common/bo/StringMetaDataBo.java
