+++ commons-server/src/main/java/com/navercorp/pinpoint/common/util/concurrent/CommonState.java
