+++ commons-server/src/test/java/com/navercorp/pinpoint/common/bo/AnnotationBoTest.java
