+++ commons-server/src/test/java/com/navercorp/pinpoint/common/bo/SpanEventBoTest.java
