+++ plugins/redis/pom.xml
@@ -23,6 +23,11 @@
             <artifactId>pinpoint-profiler</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.navercorp.pinpoint</groupId>
+            <artifactId>pinpoint-commons-server</artifactId>
+            <scope>test</scope>
+        </dependency>
 
         <dependency>
             <groupId>redis.clients</groupId>
