+++ plugins/tomcat/pom.xml
@@ -29,6 +29,11 @@
             <artifactId>pinpoint-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.navercorp.pinpoint</groupId>
+            <artifactId>pinpoint-commons-server</artifactId>
+            <scope>test</scope>
+        </dependency>
 
 
         <dependency>
